requirements.txt
hydra-core
torch
torchvision
numpy
